[KeyedService] Introduce KeyedServiceTemplatedFactory
The classes KeyedServiceFactory and RefcountedKeyedServiceFactory
implementation was almost identical except for a few types (owned
pointers, non-owned pointers, factory callback) and the name of
the shutdown method.

Introduce a templated class KeyedServiceTemplatedFactory taking
the service type as a parameter, and use helper traits template
to factor the differences. The class definition is in a .cc file
with explicit instantiation as there are only two sub-classes (it
avoids having the compiler generates code for the template in all
factories and thus should reduce the risk of regressing build-time
with this change).

This removes parallel implementation and will make simplify future
changes to KeyedServiceFactory and RefcountedKeyedServiceFactory
classes (since there won't be risk to have divergent between the
two implementations).

Bug: 342112724
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5569618
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309899}

+++ chrome/browser/profiles/profile_metrics.cc
@@ -35,7 +35,7 @@ enum class ProfileNetUserCounts {
   kMaxValue = PROFILE_DELETED
 };
 
-int GetTotalKeyedServiceCount(Profile* profile) {
+size_t GetTotalKeyedServiceCount(Profile* profile) {
   return KeyedServiceFactory::GetServicesCount(profile) +
          RefcountedKeyedServiceFactory::GetServicesCount(profile);
 }
